+++ changelog.md
@@ -31,3 +31,4 @@
 - `World::is_body_passable` has been removed. Use `PhysicalBody::passable` (returned by `World::body`) instead.
 - The `passable` field on objects is respected by `NphysicsWorld`:
   Passable objects can overlap with other objects including other passable objects.
+- All traits except `ObjectBehavior` have been sealed and can therefore no longer be implemented by other crates.